SQL Request update

+++ cart.php
@@ -7,8 +7,8 @@
         $req->execute(Array($NOM));
         $donnees = $req->fetch();
 
-        if(isset($_POST['id'])){
-            $id=$_POST['id'];
+        if(isset($_POST['id_produit'])){
+            $id=$_POST['id_produit'];
 
             if(isset($_POST['quantity'])){
                 $quantity=$_POST['quantity'];
@@ -16,9 +16,9 @@
                 try {
                     include "bdd.php";
                     if($quantity == 0){
-                        $req4 = $conn->prepare('DELETE FROM order_products WHERE product_id = '.$id);
+                        $req4 = $conn->prepare('DELETE FROM order_products, orders WHERE (order_products.product_id = '.$id_produit.' AND order_products.order_id = orders.id AND orders.user_id = ?)');
                     }else{
-                        $req4 = $conn->prepare('UPDATE order_products SET quantity = '.$quantity.' WHERE product_id = '.$id);
+                        $req4 = $conn->prepare('UPDATE order_products FROM order_products, orders SET quantity = '.$quantity.' WHERE product_id = '.$id_produit.' AND order_products.order_id = orders.id AND orders.user_id = ?)');
                     }                    
                     $req4->execute();
                 }
@@ -81,7 +81,7 @@ $conn = null;
                 <div class="price"> <?php echo $data["unit_price"]?>€ </div>
                 <form action='' method='post'>
                     <div class="quantity">Quantity: <input class="number" type="number" name="quantity" value=<?php echo $data["quantity"] ?> min="0" max="99"> </div>
-                    <input type="hidden" name="id" value=<?php echo $data["id"] ?>>
+                    <input type="hidden" name="id_produit" value=<?php echo $data["id"] ?>>
                     <input class="bouton_1" type="submit" value="Valider"/>
                     <p><?php echo $data["description"] ?></p>
                 </form>
